+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Support updating Mailchimp when a user changes their email in Firebase (PR #46)
 - Support extended syntax for configurations of Member Tags and Member Events, to allow multidimensional arrays to be unwrapped (PR #53)
+- Added configuration validation via JSON Schema (PR #54)
 
 ## Version 0.4.0